Use intptr_t to avoid assumption that integers are 32 bit.
Compiling wxHaskell on a 64bits platform fails, because there
are some assumptions in the code that assume that integers are
32bits. For example, there are a couple of assignments of a
pointer to an int.

This patch fixes these problems. Sometimes, after assigning a
pointer to an integer, the int was used as a boolean value. In
those cases, instead of returning a pointer p, I changed the
code to return p != NULL. In two other cases, a pointer is stored
inside the int and in these cases I changed the type of the
int to intptr_t, of which it is guaranteed that a pointer can be
assigned to it and recovered from it.

Requires <inttypes.h>.

darcs-hash:20061125125628-916c1-16a704ca7eb8abec1d82f25066aaa1285d9ef4ad.gz

+++ wxc/src/extra.cpp
@@ -192,7 +192,7 @@ public:
   wxInputSink( wxInputStream* input, wxEvtHandler* evtHandler, int bufferLen );
   ~wxInputSink();
 
-  int  GetId();
+  intptr_t GetId();
   void Start();
 };
 
@@ -219,9 +219,9 @@ void wxInputSink::Start()
   }
 }
 
-int wxInputSink::GetId()
+intptr_t wxInputSink::GetId()
 {
-  return (int)m_input;
+  return (intptr_t)m_input;
 }
 
 wxThread::ExitCode wxInputSink::Entry()
